feat: remove ruleTabReducer
