--- ui/react/pages/RuleProfile/Tabs/Limit/actionTypes.js
@@ -1,5 +1,0 @@
-export const CHANGE_DROPDOWN_FIELD = Symbol('CHANGE_DROPDOWN_FIELD');
-export const CHANGE_INPUT_FIELD = Symbol('CHANGE_INPUT_FIELD');
-export const ADD_LIMIT = Symbol('ADD_LIMIT');
-export const REMOVE_LIMIT = Symbol('REMOVE_LIMIT');
-export const SET_LIMIT_FIELD = Symbol('SET_LIMIT_FIELD');
