+++ ui/react/pages/RuleProfile/Tabs/Source/index.js
@@ -6,8 +6,8 @@ import MultiSelectBubble from 'ut-front-react/components/MultiSelectBubble';
 import Dropdown from 'ut-front-react/components/Input/Dropdown';
 import Input from 'ut-front-react/components/Input';
 import style from '../style.css';
-import * as actions from './actions';
-
+import * as actions from '../../actions';
+const destinationProp = 'source';
 const propTypes = {
     actions: PropTypes.object,
     countries: PropTypes.array,
@@ -46,7 +46,12 @@ class SourceTab extends Component {
 
     getPropetyRowsBody() {
         const { properties } = this.props.fieldValues;
-        const { removeProperty, setPropertyField } = this.props.actions;
+        let removeProperty = (index) => {
+            this.props.actions.removeProperty(index, destinationProp);
+        };
+        let setPropertyField = (index, key, value) => {
+            this.props.actions.setPropertyField(index, key, value, destinationProp);
+        };
         return properties.map((prop, index) => {
             return (
                 <tr key={`${index}`}>
@@ -73,7 +78,9 @@ class SourceTab extends Component {
     }
 
     renderPropertyTable() {
-        const { addProperty } = this.props.actions;
+        let addProperty = () => {
+            this.props.actions.addProperty(destinationProp);
+        };
         return (
             <div className={style.propertyTable}>
                 <table className={style.dataGridTable}>
@@ -104,11 +111,12 @@ class SourceTab extends Component {
             fieldValues
         } = this.props;
 
-        const {
-            changeMultiSelectField,
-            changeDropdownField
-        } = this.props.actions;
-
+        let changeMultiSelectField = (field, value) => {
+            this.props.actions.changeMultiSelectField(field, value, destinationProp);
+        };
+        let changeDropdownField = (field, value) => {
+            this.props.actions.changeDropdownField(field, value, destinationProp);
+        };
         return (
             <div>
                 <div className={style.inputWrapper}>
@@ -199,12 +207,12 @@ SourceTab.defaultProps = defaultProps;
 
 const mapStateToProps = (state, ownProps) => {
     return {
-        countries: state.ruleTabReducer.getIn(['nomenclatures', 'country']).toJS(),
-        regions: state.ruleTabReducer.getIn(['nomenclatures', 'region']).toJS(),
-        cities: state.ruleTabReducer.getIn(['nomenclatures', 'city']).toJS(),
-        organizations: state.ruleTabReducer.getIn(['nomenclatures', 'organization']).toJS(),
-        products: state.ruleTabReducer.getIn(['nomenclatures', 'cardProduct']).toJS(),
-        fieldValues: state.ruleSourceTabReducer.get('fields').toJS()
+        countries: state.ruleProfileReducer.getIn(['nomenclatures', 'country']).toJS(),
+        regions: state.ruleProfileReducer.getIn(['nomenclatures', 'region']).toJS(),
+        cities: state.ruleProfileReducer.getIn(['nomenclatures', 'city']).toJS(),
+        organizations: state.ruleProfileReducer.getIn(['nomenclatures', 'organization']).toJS(),
+        products: state.ruleProfileReducer.getIn(['nomenclatures', 'cardProduct']).toJS(),
+        fieldValues: state.ruleProfileReducer.get(destinationProp).toJS()
     };
 };
