--- ui/react/pages/RuleProfile/Tabs/Split/actionTypes.js
@@ -1,16 +1,0 @@
-export const CHANGE_MULTISELECT_FIELD = Symbol('CHANGE_MULTISELECT_FIELD');
-export const CHANGE_DROPDOWN_FIELD = Symbol('CHANGE_DROPDOWN_FIELD');
-export const ADD_ASSIGNMENT = Symbol('ADD_ASSIGNMENT');
-export const REMOVE_ASSIGNMENT = Symbol('REMOVE_ASSIGNMENT');
-export const SET_ASSIGNMENT_FIELD = Symbol('SET_ASSIGNMENT_FIELD');
-export const CHANGE_INPUT_FIELD = Symbol('CHANGE_INPUT_FIELD');
-export const ADD_CUMULATIVE = Symbol('ADD_CUMULATIVE');
-export const REMOVE_CUMULATIVE = Symbol('REMOVE_CUMULATIVE');
-export const SET_CUMULATIVE_FIELD = Symbol('SET_CUMULATIVE_FIELD');
-
-export const ADD_CUMULATIVE_RANGE = Symbol('ADD_CUMULATIVE_RANGE');
-export const REMOVE_CUMULATIVE_RANGE = Symbol('REMOVE_CUMULATIVE_RANGE');
-export const SET_CUMULATIVE_RANGE_FIELD = Symbol('SET_CUMULATIVE_RANGE_FIELD');
-
-export const ADD_SPLIT = Symbol('ADD_SPLIT');
-export const REMOVE_SPLIT = Symbol('REMOVE_SPLIT');
