+++ ui/react/pages/RuleProfile/actionTypes.js
@@ -2,3 +2,30 @@
 export const FETCH_NOMENCLATURES = Symbol('FETCH_NOMENCLATURES');
 export const CREATE_RULE = Symbol('CREATE_RULE');
 export const RESET_RULE_STATE = Symbol('RESET_RULE_STATE');
+
+// tab common actions
+export const CHANGE_MULTISELECT_FIELD = Symbol('CHANGE_MULTISELECT_FIELD');
+export const CHANGE_DROPDOWN_FIELD = Symbol('CHANGE_DROPDOWN_FIELD');
+export const ADD_PROPERTY = Symbol('ADD_PROPERTY');
+export const REMOVE_PROPERTY = Symbol('REMOVE_PROPERTY');
+export const SET_PROPERTY_FIELD = Symbol('SET_PROPERTY_FIELD');
+export const CHANGE_INPUT_FIELD = Symbol('CHANGE_INPUT_FIELD');
+// tab limit actions
+export const ADD_LIMIT = Symbol('ADD_LIMIT');
+export const REMOVE_LIMIT = Symbol('REMOVE_LIMIT');
+export const SET_LIMIT_FIELD = Symbol('SET_LIMIT_FIELD');
+// tab split actions
+export const CHANGE_SPLIT_MULTISELECT_FIELD = Symbol('CHANGE_SPLIT_MULTISELECT_FIELD');
+// export const CHANGE_SPLIT_DROPDOWN_FIELD = Symbol('CHANGE_SPLIT_DROPDOWN_FIELD');
+export const CHANGE_SPLIT_INPUT_FIELD = Symbol('CHANGE_SPLIT_INPUT_FIELD');
+export const ADD_ASSIGNMENT = Symbol('ADD_ASSIGNMENT');
+export const REMOVE_ASSIGNMENT = Symbol('REMOVE_ASSIGNMENT');
+export const SET_ASSIGNMENT_FIELD = Symbol('SET_ASSIGNMENT_FIELD');
+export const ADD_CUMULATIVE = Symbol('ADD_CUMULATIVE');
+export const REMOVE_CUMULATIVE = Symbol('REMOVE_CUMULATIVE');
+export const SET_CUMULATIVE_FIELD = Symbol('SET_CUMULATIVE_FIELD');
+export const ADD_CUMULATIVE_RANGE = Symbol('ADD_CUMULATIVE_RANGE');
+export const REMOVE_CUMULATIVE_RANGE = Symbol('REMOVE_CUMULATIVE_RANGE');
+export const SET_CUMULATIVE_RANGE_FIELD = Symbol('SET_CUMULATIVE_RANGE_FIELD');
+export const ADD_SPLIT = Symbol('ADD_SPLIT');
+export const REMOVE_SPLIT = Symbol('REMOVE_SPLIT');
